Phase 2b step A: rename lcm/grids/ → lcm/_grids/
`lcm/grids/` is internal grid infrastructure (ABCs, validators,
coordinate helpers, the leaf classes whose user-facing copies live
in `lcm/api/grids.py`). The leading underscore signals "private —
don't import from user code"; users keep reaching for grid classes
through `from lcm import LinSpacedGrid` or `from lcm.api.grids
import LinSpacedGrid`.

Pure rename via `git mv` to preserve blame. Sed-rewrite of `from
lcm.grids` / `import lcm.grids` / `lcm.grids.` across src and tests.
Update docstring references in `api/grids.py` and `api/categorical.py`.

Step A of `Phase 2 — api Reorganisation.md`'s deferred internal
restructure.
