+++ src/lcm/_grids/base.py
