+++ src/lcm/_grids/categorical.py
