+++ src/lcm/api/categorical.py
@@ -2,12 +2,10 @@
 
 Used to declare a category enumeration class whose fields become the
 `DiscreteGrid` codes for a discrete state / action. The decorator and its
-supporting validators live in `lcm.grids.categorical` (private
-implementation detail today; will consolidate under `lcm.api` in a
-follow-up).
+supporting validators live in `lcm._grids.categorical`.
 """
 
-from lcm.grids.categorical import categorical, validate_category_class
+from lcm._grids.categorical import categorical, validate_category_class
 
 __all__ = [
     "categorical",
