+++ src/lcm/api/regime.py
@@ -13,8 +13,8 @@ from beartype import beartype
 from dags.tree import QNAME_DELIMITER
 
 from lcm._beartype_conf import REGIME_CONF
+from lcm._grids import DiscreteGrid, Grid
 from lcm.exceptions import RegimeInitializationError, format_messages
-from lcm.grids import DiscreteGrid, Grid
 from lcm.shocks._base import _ShockGrid
 from lcm.typing import (
     ActionName,
