+++ src/lcm/params/regime_template.py
@@ -3,10 +3,10 @@ from types import MappingProxyType
 import dags.tree as dt
 from dags.tree import tree_path_from_qname
 
+from lcm._grids import IrregSpacedGrid
 from lcm.api.regime import Regime as UserRegime
 from lcm.api.regime import SolveSimulateFunctionPair
 from lcm.exceptions import InvalidNameError
-from lcm.grids import IrregSpacedGrid
 from lcm.regime_building.transitions import collect_state_transitions
 from lcm.shocks import _ShockGrid
 from lcm.typing import (
