+++ src/lcm/regime_building/next_state.py
@@ -7,8 +7,8 @@ import jax
 from dags import concatenate_functions, with_signature
 from dags.tree import qname_from_tree_path
 
+from lcm._grids import Grid
 from lcm.engine import Variables
-from lcm.grids import Grid
 from lcm.shocks import _ShockGrid
 from lcm.shocks.ar1 import _ShockGridAR1
 from lcm.shocks.iid import _ShockGridIID
