+++ src/lcm/regime_building/transitions.py
@@ -5,9 +5,9 @@ from typing import TypeAliasType
 
 from dags.tree import QNAME_DELIMITER
 
+from lcm._grids import DiscreteGrid, Grid
 from lcm.api.regime import _IdentityTransition
 from lcm.exceptions import RegimeInitializationError
-from lcm.grids import DiscreteGrid, Grid
 from lcm.shocks._base import _ShockGrid
 from lcm.typing import (
     ContinuousState,
