+++ tests/mock_regime.py
@@ -1,9 +1,9 @@
 from types import MappingProxyType
 from typing import Literal, cast
 
+from lcm._grids import Grid
 from lcm.api.regime import Regime as UserRegime
 from lcm.api.regime import SolveSimulateFunctionPair, _default_H
-from lcm.grids import Grid
 from lcm.regime_building.transitions import collect_state_transitions
 from lcm.typing import UserFunction
