+++ tests/regime_building/test_create_regime_params_template.py
@@ -1,5 +1,5 @@
+from lcm._grids import DiscreteGrid
 from lcm.api.regime import SolveSimulateFunctionPair
-from lcm.grids import DiscreteGrid
 from lcm.params.regime_template import (
     create_regime_params_template,
 )
