+++ tests/regime_building/test_regime_processing.py
@@ -6,10 +6,10 @@ import pytest
 from numpy.testing import assert_array_equal
 
 from lcm import categorical
+from lcm._grids import DiscreteGrid, LinSpacedGrid
 from lcm.api.ages import AgeGrid
 from lcm.api.regime import Regime as UserRegime
 from lcm.engine import Regime, VariableInfo, Variables
-from lcm.grids import DiscreteGrid, LinSpacedGrid
 from lcm.regime_building.processing import (
     _rename_params_to_qnames,
     process_regimes,
