+++ tests/test_next_state.py
@@ -2,9 +2,9 @@ from types import MappingProxyType
 
 import jax.numpy as jnp
 
+from lcm._grids import DiscreteGrid, categorical
 from lcm.api.ages import AgeGrid
 from lcm.engine import VariableInfo, Variables
-from lcm.grids import DiscreteGrid, categorical
 from lcm.regime_building import process_regimes
 from lcm.regime_building.next_state import (
     _create_discrete_stochastic_next_func,
