+++ tests/test_runtime_params.py
@@ -5,8 +5,8 @@ import pytest
 from numpy.testing import assert_array_almost_equal as aaae
 
 from lcm import AgeGrid, LinSpacedGrid, Model, categorical
+from lcm._grids import IrregSpacedGrid
 from lcm.api.regime import Regime as UserRegime
-from lcm.grids import IrregSpacedGrid
 from lcm.typing import ContinuousAction, ContinuousState, FloatND, ScalarInt
