+++ tests/test_single_feasible_action.py
@@ -25,9 +25,9 @@ import jax.numpy as jnp
 import pytest
 
 from lcm import AgeGrid, DiscreteGrid, LinSpacedGrid, Model, categorical
+from lcm._grids import IrregSpacedGrid
+from lcm._grids.coordinates import get_irreg_coordinate
 from lcm.api.regime import Regime as UserRegime
-from lcm.grids import IrregSpacedGrid
-from lcm.grids.coordinates import get_irreg_coordinate
 from lcm.regime_building.ndimage import map_coordinates
 from lcm.typing import (
     ContinuousAction,
